requirements.txt
# AI Gameplay Bot - Python Dependencies
# Python 3.8+ required
# Version constraints: compatibility ranges with flexible upper bounds

# Deep Learning & Neural Networks
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0
transformers>=4.30.0,<5.0.0
numpy>=1.24.0,<2.0.0

# Computer Vision
opencv-python>=4.8.0,<5.0.0
Pillow>=10.0.0,<11.0.0

# Data Processing & Analysis
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
scipy>=1.11.0,<2.0.0

# Web Framework & API
Flask>=2.3.0,<4.0.0
Flask-CORS>=4.0.0,<5.0.0
requests>=2.31.0,<3.0.0
gunicorn>=21.2.0,<22.0.0

# Reinforcement Learning
#gym>=0.26.0,<1.0.0
#stable-baselines3>=2.0.0,<3.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
tqdm>=4.65.0,<5.0.0
pyyaml>=6.0.0,<7.0.0
colorama>=0.4.6,<1.0.0

# Input Recording & Screen Capture
pynput>=1.7.0,<2.0.0
mss>=9.0.0,<10.0.0

# Monitoring & Logging
prometheus-client>=0.17.0,<1.0.0
psutil>=5.9.0,<6.0.0

# Testing
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-mock>=3.11.0,<4.0.0
pytest-asyncio>=0.21.0,<1.0.0

# Code Quality
black>=23.7.0,<25.0.0
flake8>=6.0.0,<7.0.0
pylint>=2.17.0,<4.0.0
mypy>=1.4.0,<2.0.0

# Development
jupyter>=1.0.0,<2.0.0
ipython>=8.14.0,<9.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# Stream URL resolution (YouTube/Twitch)
yt-dlp>=2024.0.0,<2027.0.0
